+++ front/src/Home.tsx
@@ -4,12 +4,12 @@ import TweetForm from './TweetForm';
 import useCallApi from './utils/api';
 import dayjs from 'dayjs';
 import { TweetType } from './utils/types';
+import ReplyButton from './ReplyButton';
 import FavButton from './FavButton';
 import './Home.css';
 
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
-import ChatBubbleOutlineIcon from '@mui/icons-material/ChatBubbleOutline';
 import CompareArrowsIcon from '@mui/icons-material/CompareArrows';
 import IosShareIcon from '@mui/icons-material/IosShare';
 import Divider from '@mui/material/Divider';
@@ -20,29 +20,10 @@ import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import Modal from '@mui/material/Modal';
-import Box from '@mui/material/Box';
 
 const Home: React.FC = () => {
     const [tweetsList, setTweets] = useState<Array<TweetType>>([]);
     const callApi = useCallApi();
-    const [modalTweet, setModalTweet] = useState<TweetType>();
-
-    const [open, setOpen] = React.useState(false);
-    const handleOpen = () => setOpen(true);
-    const handleClose = () => setOpen(false);
-
-    const style = {
-        position: 'absolute' as 'absolute',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%, -50%)',
-        width: 400,
-        bgcolor: 'background.paper',
-        border: '2px solid #000',
-        boxShadow: 24,
-        p: 4,
-    };
 
     useEffect(() => {
         getTweets()?.then(setTweets);
@@ -90,11 +71,6 @@ const Home: React.FC = () => {
         getTweets()?.then(setTweets);
     }
 
-    function setModalTweetAndHandleOpen(modalTweet: TweetType): void {
-        setModalTweet(modalTweet);
-        handleOpen();
-    }
-
     function formatDate(createdAt: Date): string {
         const now: dayjs.Dayjs = dayjs();
         if (dayjs(createdAt).isBefore(now.subtract(1, 'd'))) {
@@ -137,40 +113,11 @@ const Home: React.FC = () => {
                                     <Link to={'/tweet/' + tweet.id}>{tweet.content}</Link>
                                     <br />
                                     <ListItemIcon>
-                                        <Button onClick={() => setModalTweetAndHandleOpen(tweet)}>
-                                            <ChatBubbleOutlineIcon />
-                                        </Button>
-                                        <Modal
-                                            open={open}
-                                            onClose={handleClose}
-                                            aria-labelledby="modal-modal-title"
-                                            aria-describedby="modal-modal-description"
-                                            className="Modal"
-                                        >
-                                            <Box sx={style}>
-                                                {modalTweet ? (
-                                                    <>
-                                                        <Typography id="modal-modal-title" variant="h6" component="h2">
-                                                            <Avatar
-                                                                alt={modalTweet.user.username}
-                                                                src={modalTweet.user.icon}
-                                                            />
-                                                            {modalTweet.user.username}
-                                                            <br />
-                                                            {modalTweet.content}
-                                                        </Typography>
-                                                        <TweetForm
-                                                            getAndSetTweets={() => getAndSetTweets()}
-                                                            caller={'reply'}
-                                                            replyTo={modalTweet.id}
-                                                        />
-                                                    </>
-                                                ) : (
-                                                    'ツイートを取得できませんでした'
-                                                )}
-                                            </Box>
-                                        </Modal>
-                                        {tweet.numberOfReply}
+                                        <ReplyButton
+                                            tweet={tweet}
+                                            numberOfReply={tweet.numberOfReply}
+                                            getAndSetTweets={() => getAndSetTweets()}
+                                        />
                                         <Button variant="text">
                                             <CompareArrowsIcon />
                                         </Button>
